add some bullets
三方向と八方向に飛んでいく弾を実装しました。現状だとポイント・経験値による解放に関係なく、撃ててしまいます。

+++ Assets/Scripts/Player/Player.cs
@@ -229,7 +229,7 @@ public class Player : MonoBehaviour, IDamageable
         _shootingCoroutine = null;
     }
 
-    public int GetPlayerHP()
+    public float GetPlayerHP()
     {
         return _playerHP;
     }
